+++ network/rpc/handlers_chain.go
@@ -267,7 +267,7 @@ func (server *CommonChainRPC) verifyChain(level, depth int32) error {
 		// Level 1 does basic BlockChain sanity checks.
 		if level > 0 {
 			err := chaindata.CheckBlockSanity(block,
-				server.chainProvider.ChainParams.PowParams.PowLimit, server.chainProvider.TimeSource)
+				server.chainProvider.ChainParams, server.chainProvider.TimeSource)
 			if err != nil {
 				server.Log.Error().Msgf("Verify is unable to validate "+
 					"block at hash %v height %d: %v",