+++ node/blockchain/chain.go
@@ -568,7 +568,7 @@ func (b *BlockChain) calcSequenceLock(node blocknode.IBlockNode, tx *jaxutil.Tx,
 	// can be included within a block at any given height or time.
 	mTx := tx.MsgTx()
 
-	sequenceLockActive := mTx.CleanVersion() == wire.TxVerTimeLock && csvSoftforkActive
+	sequenceLockActive := mTx.Version == wire.TxVerTimeLock && csvSoftforkActive
 	if !sequenceLockActive || chaindata.IsCoinBase(tx) {
 		return sequenceLock, nil
 	}