+++ node/mempool/policy.go
@@ -302,10 +302,9 @@ func checkTransactionStandard(tx *jaxutil.Tx, height int32,
 
 	// The transaction must be a currently supported version.
 	msgTx := tx.MsgTx()
-	if msgTx.CleanVersion() > maxTxVersion || msgTx.CleanVersion() < wire.TxVerRegular {
+	if msgTx.Version > maxTxVersion || msgTx.Version < wire.TxVerRegular {
 		str := fmt.Sprintf("transaction version %d is not in the "+
-			"valid range of %d-%d", msgTx.CleanVersion(), wire.TxVerRegular,
-			maxTxVersion)
+			"valid range of %d-%d", msgTx.Version, wire.TxVerRegular, maxTxVersion)
 		return txRuleError(wire.RejectNonstandard, str)
 	}