+++ types/blocknode/block_node.shard.go
@@ -19,8 +19,6 @@ const (
 	shardMedianTimeBlocks = 11
 )
 
-var zeroHash chainhash.Hash
-
 // ShardBlockNode represents a block within the block chain and is primarily used to
 // aid in selecting the best chain to be the main chain.  The main chain is
 // stored into the block database.
@@ -32,53 +30,19 @@ type ShardBlockNode struct {
 	// hundreds of thousands of these in memory, so a few extra bytes of
 	// padding adds up.
 
-	parent   IBlockNode     // parent is the parent block for this node.
-	hash     chainhash.Hash // hash is the double sha 256 of the block.
-	workSum  *big.Int       // workSum is the total amount of work in the chain up to and including this node.
-	height   int32          // height is the position in the block chain.
-	serialID int64          // serialID is the absolute unique id of current block.
+	parent    IBlockNode     // parent is the parent block for this node.
+	hash      chainhash.Hash // hash is the double sha 256 of the block.
+	workSum   *big.Int       // workSum is the total amount of work in the chain up to and including this node.
+	height    int32          // height is the position in the block chain.
+	serialID  int64          // serialID is the absolute unique id of current block.
+	timestamp int64
 
-	// todo(mike): review this, might be make sense to store raw bytes or wire obj instead
-	// Some fields from block headers to aid in best chain selection and
-	// reconstructing headers from memory.  These must be treated as
-	// immutable and are intentionally ordered to avoid padding on 64-bit
-	// platforms.
-	version           int32
-	bits              uint32
-	timestamp         int64
-	merkleRoot        chainhash.Hash
-	mergeMiningNumber uint32
-	bcHeader          beaconBlockNodeFields
-
+	header wire.BlockHeader
 	// status is a bitfield representing the validation state of the block. The
 	// status field, unlike the other fields, may be written to and so should
 	// only be accessed using the concurrent-safe NodeStatus method on
 	// blockIndex once the node has been added to the global index.
 	status BlockStatus
-}
-
-type beaconBlockNodeFields struct {
-	hash    chainhash.Hash // hash is the double sha 256 of the block.
-	workSum *big.Int       // workSum is the total amount of work in the chain up to and including this node.
-
-	// todo(mike): review this, might be make sense to store raw bytes or wire obj instead
-
-	// Some fields from block headers to aid in best chain selection and
-	// reconstructing headers from memory.  These must be treated as
-	// immutable and are intentionally ordered to avoid padding on 64-bit
-	// platforms.
-	version       int32
-	bits          uint32
-	timestamp     int64
-	prevHash      chainhash.Hash
-	merkleRoot    chainhash.Hash
-	mmrRoot       chainhash.Hash
-	treeEncoding  []uint8          // Encoding of the Merge-mining tree
-	shards        uint32           // A number of shards at moment, when block was mined.
-	k             uint32           // k is inflation-fix coefficient for current mining epoch.
-	voteK         uint32           // voteK is a proposed inflation-fix coefficient for next mining epoch.
-	btcAux        wire.BTCBlockAux // A bitcoin header auxiliary, required by merge mining protocol.
-	bcCoinbaseAux wire.CoinbaseAux // A beacon coinbase tx auxiliary, required by merge mining protocol.
 }
 
 // NewShardBlockNode returns a new block node for the given block ShardHeader and parent
@@ -93,33 +57,12 @@ func NewShardBlockNode(blockHeader wire.BlockHeader, parent IBlockNode) *ShardBl
 // This function is NOT safe for concurrent access.  It must only be called when
 // initially creating a node.
 func initShardBlockNode(blockHeader wire.BlockHeader, parent IBlockNode) *ShardBlockNode {
-	shardHeader, _ := blockHeader.(*wire.ShardHeader)
-	beaconHeader := blockHeader.BeaconHeader()
 	node := &ShardBlockNode{
-		hash:              shardHeader.BlockHash(),
-		workSum:           pow.CalcWork(shardHeader.Bits()),
-		version:           int32(shardHeader.Version()),
-		bits:              shardHeader.Bits(),
-		timestamp:         shardHeader.Timestamp().Unix(),
-		merkleRoot:        shardHeader.MerkleRoot(),
-		mergeMiningNumber: shardHeader.MergeMiningNumber(),
-
-		bcHeader: beaconBlockNodeFields{
-			hash:          beaconHeader.BlockHash(),
-			workSum:       pow.CalcWork(beaconHeader.Bits()),
-			version:       int32(beaconHeader.Version()),
-			bits:          beaconHeader.Bits(),
-			timestamp:     beaconHeader.Timestamp().Unix(),
-			merkleRoot:    beaconHeader.MerkleRoot(),
-			mmrRoot:       beaconHeader.MergeMiningRoot(),
-			shards:        beaconHeader.Shards(),
-			treeEncoding:  beaconHeader.MergedMiningTree(),
-			prevHash:      beaconHeader.PrevBlock(),
-			k:             beaconHeader.K(),
-			voteK:         beaconHeader.VoteK(),
-			btcAux:        *beaconHeader.BTCAux().Copy(),
-			bcCoinbaseAux: *shardHeader.CoinbaseAux.Copy(),
-		},
+		hash:      blockHeader.BlockHash(),
+		workSum:   pow.CalcWork(blockHeader.Bits()),
+		timestamp: blockHeader.Timestamp().Unix(),
+		// header:    blockHeader.Copy(),
+		header: blockHeader,
 	}
 	if parent != nil {
 		node.parent = parent
@@ -132,11 +75,11 @@ func initShardBlockNode(blockHeader wire.BlockHeader, parent IBlockNode) *ShardB
 }
 
 func (node *ShardBlockNode) GetHash() chainhash.Hash      { return node.hash }
-func (node *ShardBlockNode) Version() int32               { return node.version }
+func (node *ShardBlockNode) Version() int32               { return node.header.Version().Version() }
 func (node *ShardBlockNode) Height() int32                { return node.height }
 func (node *ShardBlockNode) SerialID() int64              { return node.serialID }
-func (node *ShardBlockNode) Bits() uint32                 { return node.bits }
-func (node *ShardBlockNode) K() uint32                    { return node.bcHeader.k }
+func (node *ShardBlockNode) Bits() uint32                 { return node.header.Bits() }
+func (node *ShardBlockNode) K() uint32                    { return node.header.K() }
 func (node *ShardBlockNode) Parent() IBlockNode           { return node.parent }
 func (node *ShardBlockNode) WorkSum() *big.Int            { return node.workSum }
 func (node *ShardBlockNode) Timestamp() int64             { return node.timestamp }
@@ -148,40 +91,8 @@ func (node *ShardBlockNode) NewHeader() wire.BlockHeader  { return wire.EmptySha
 //
 // This function is safe for concurrent access.
 func (node *ShardBlockNode) Header() wire.BlockHeader {
-	// No lock is needed because all accessed fields are immutable.
-	prevHash := &zeroHash
-	if node.parent != nil {
-		h := node.parent.GetHash()
-		prevHash = &h
-	}
-
-	beaconHeader := wire.NewBeaconBlockHeader(
-		wire.BVersion(node.bcHeader.version),
-		node.bcHeader.prevHash,
-		node.bcHeader.merkleRoot,
-		node.bcHeader.mmrRoot,
-		time.Unix(node.bcHeader.timestamp, 0),
-		node.bcHeader.bits,
-		node.bcHeader.btcAux.Nonce)
-
-	beaconHeader.SetMergedMiningTree(node.bcHeader.treeEncoding)
-	beaconHeader.SetK(node.bcHeader.k)
-	beaconHeader.SetVoteK(node.bcHeader.voteK)
-	beaconHeader.BeaconHeader().SetBTCAux(node.bcHeader.btcAux)
-
-	beaconHeader.SetShards(node.bcHeader.shards)
-	beaconHeader.SetMergedMiningTree(node.bcHeader.treeEncoding)
-
-	header := wire.NewShardBlockHeader(
-		*prevHash,
-		node.merkleRoot,
-		time.Unix(node.timestamp, 0),
-		node.bits,
-		*beaconHeader,
-		node.bcHeader.bcCoinbaseAux,
-	)
-	header.SetMergeMiningNumber(node.mergeMiningNumber)
-	return header
+	return node.header
+	// return node.header.Copy()
 }
 
 // Ancestor returns the ancestor block node at the provided height by following
@@ -201,10 +112,6 @@ func (node *ShardBlockNode) Ancestor(height int32) IBlockNode {
 	}
 
 	return n
-}
-
-func (node *ShardBlockNode) SetBits(value uint32) {
-	node.bits = value
 }
 
 // RelativeAncestor returns the ancestor block node a relative 'distance' blocks