+++ types/chaincfg/cfg_testnet.go
@@ -1,4 +1,5 @@
-// Copyright (c) 2020 The JaxNetwork developers
+// Copyright (c) 2014-2016 The btcsuite developers
+// Copyright (c) 2020-2021 The JAX.Network developers
 // Use of this source code is governed by an ISC
 // license that can be found in the LICENSE file.
 
@@ -50,6 +51,7 @@ var TestNet3Params = Params{
 		ReduceMinDifficulty:      false,
 		MinDiffReductionTime:     time.Second * 30, // TargetTimePerBlock * 2
 		GenerateSupported:        true,
+		HashSorting:              true,
 	},
 
 	// Checkpoints ordered from oldest to newest.